Restructure project and move cli commands to root level.
