+++ clients/context.py
@@ -2,6 +2,6 @@ import sys
 sys.path.insert(0, "..")
 
 
-from sameproject.objects.step import Step
-from sameproject.program.compile import notebook_processing
+from sameproject.data.step import Step
+from sameproject.ops import notebooks
 from sameproject.backends.durable_functions.constants import EXECUTE_WORKFLOW_ACTIVITY_NAME
