+++ jupyter/kernel/context.py
@@ -4,4 +4,4 @@ sys.path.insert(0, "..")
 
 
 from clients.durable_functions_client import DurableFunctionsClient
-from sameproject.objects.step import Step
+from sameproject.data.step import Step
