+++ sameproject/backends/durable_functions/execute_step/context.py
@@ -4,7 +4,7 @@ sys.path.insert(0, "../..")
 sys.path.insert(0, "..")
 
 
-from sameproject.objects.step import Step
+from sameproject.data.step import Step
 from common import serialization_utils
 from common import code_executor
 from common import exception_utils
