+++ sameproject/backends/durable_functions/execute_steps_workflow/context.py
@@ -4,4 +4,4 @@ sys.path.insert(0, "../..")
 sys.path.insert(0, "..")
 
 
-from sameproject.objects.step import Step
+from sameproject.data.step import Step
