+++ sameproject/cli/__init__.py
@@ -1,0 +1,2 @@
+from .version import version
+from .run import run
