+++ sameproject/cli/version.py
@@ -1,11 +1,13 @@
-import click
 import pkg_resources
+import click
 
 
 @click.command()
 def version():
-    """Prints the versions for the CLI"""
+    """Prints the version of this build."""
+
     try:
         click.echo(pkg_resources.get_distribution("sameproject").version)
     except Exception as e:
         click.echo(f"Unknown version (error getting version: {e})")
+
