+++ sameproject/data/__init__.py
@@ -1,2 +1,4 @@
 """Contains objects that can be understood by and trasnferred between the SAME CLI and SAME backends.
 """
+from .step import Step
+from .json_serializable_object import JSONSerializableObject
