+++ sameproject/data/json_serializable_object.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 from typing import Any
 from abc import ABC
-
 import json
