+++ sameproject/ops/__init__.py
@@ -1,0 +1,4 @@
+"""
+Contains core SAME functionality, such as code for processing Jupyter
+notebooks and various other helper functions.
+"""
