--- sameproject/program/__init__.py
@@ -1,4 +1,0 @@
-import sys
-from pathlib import Path
-
-sys.path.append(str(Path(__file__).parent.parent.parent.parent))
