--- sameproject/program/compile/context.py
@@ -1,5 +1,0 @@
-import sys
-
-import pathlib
-
-from sameproject.objects.step import Step
