+++ test/cli/test_compile.py
@@ -1,12 +1,12 @@
-from click.testing import CliRunner
-import pytest
-from pathlib import Path
-from sameproject.same_config import SameConfig
-from sameproject.program.commands import run
-from sameproject.program.compile import notebook_processing
-import logging
 from sameproject.backends.executor import render as template_render
+from sameproject.ops import notebooks
 from sameproject.backends.executor import deploy
+from sameproject.data.same_config import SameConfig
+from click.testing import CliRunner
+from sameproject.cli import run
+from pathlib import Path
+import logging
+import pytest
 
 same_config_file_path = "test/testdata/generic_notebook/same.yaml"
 
@@ -63,14 +63,14 @@ def test_same_program_compile_e2e():
 
 
 def test_get_notebook_path(same_config):
-    assert "sample_notebook.ipynb" in notebook_processing.get_notebook_path(Path(same_config_file_path).parent, same_config)
+    assert "sample_notebook.ipynb" in notebooks.get_notebook_path(Path(same_config_file_path).parent, same_config)
 
 
 def test_bad_notebook_path(caplog):
     bad_path_string = "BAD_PATH"
     with pytest.raises(SystemExit) as e:
         with caplog.at_level(logging.FATAL):
-            notebook_processing.read_notebook(bad_path_string)
+            notebooks.read_notebook(bad_path_string)
             assert bad_path_string in caplog.text
     assert e.type == SystemExit
     assert e.value.code == 1
@@ -78,23 +78,23 @@ def test_bad_notebook_path(caplog):
 
 @pytest.mark.parametrize("test_name, notebook_path, number_of_steps, number_of_total_cells", test_notebooks, ids=[p[0] for p in test_notebooks])
 def test_parse_notebook(test_name, notebook_path, number_of_steps, number_of_total_cells):
-    notebook_dict = notebook_processing.read_notebook(notebook_path)
+    notebook_dict = notebooks.read_notebook(notebook_path)
     assert notebook_dict.get("cells", None) is not None
     assert (
         len(notebook_dict["cells"]) == number_of_total_cells
     ), f"{test_name} did not get number of expected cells - expected: {number_of_total_cells}, actual: {len(notebook_dict['cells'])}"
 
-    steps = notebook_processing.get_steps(notebook_dict)
+    steps = notebooks.get_steps(notebook_dict)
     assert len(steps) == number_of_steps, f"{test_name} did not get number of expected steps - expected: {number_of_steps}, actual: {len(steps)}"
 
 
 @pytest.mark.parametrize("test_name, error_string", magic_strings_to_detect, ids=[p[0] for p in magic_strings_to_detect])
 def test_detect_bad_python_strings(caplog, test_name, error_string):
     notebook_path = "test/testdata/notebook_edge_cases/bad_python_lines.ipynb"
-    notebook_dict = notebook_processing.read_notebook(notebook_path)
+    notebook_dict = notebooks.read_notebook(notebook_path)
     with pytest.raises(SyntaxError) as e:
         with caplog.at_level(logging.INFO):
-            notebook_processing.get_steps(notebook_dict)
+            notebooks.get_steps(notebook_dict)
             assert error_string in caplog.text
     assert e.type == SyntaxError
 
@@ -105,13 +105,13 @@ def test_e2e_full_notebook():
     number_of_total_cells = 13
     test_name = "E2E 'sample_notebook'"
 
-    notebook_dict = notebook_processing.read_notebook(notebook_path)
+    notebook_dict = notebooks.read_notebook(notebook_path)
     assert notebook_dict.get("cells", None) is not None
     assert (
         len(notebook_dict["cells"]) == number_of_total_cells
     ), f"{test_name} did not get number of expected cells - expected: {number_of_total_cells}, actual: {len(notebook_dict['cells'])}"
 
-    steps = notebook_processing.get_steps(notebook_dict)
+    steps = notebooks.get_steps(notebook_dict)
     assert len(steps) == number_of_steps, f"{test_name} did not get number of expected steps - expected: {number_of_steps}, actual: {len(steps)}"
 
     assert "tensorflow" in steps["same_step_000"].packages_to_install, f"Packages to install expected to contain 'tensorflow'. Actual: {steps['same_step_000'].packages_to_install}"
@@ -131,7 +131,7 @@ def test_kubeflow_secrets(mocker, tmpdir, same_config):
 
     # No Secrets
     with open(same_config_file_path, "rb") as f:
-        notebook_processing.compile(f, "kubeflow")
+        notebooks.compile(f, "kubeflow")
 
     # Root file is the 4rd call zero indexed (there should be a cleaner way to do this - super fragile)
     root_file_content = Path.write_text.call_args_list[3][0][0]
@@ -150,7 +150,7 @@ def test_kubeflow_secrets(mocker, tmpdir, same_config):
 
     multienv_same_file = "test/testdata/multienv_notebook/same.yaml"
     with open(multienv_same_file, "rb") as f:
-        notebook_processing.compile(f, "kubeflow", secret_dict)
+        notebooks.compile(f, "kubeflow", secret_dict)
 
     # Root file is the 7th call zero indexed (there should be a cleaner way to do this - super fragile)
     root_file_content = Path.write_text.call_args_list[6][0][0]
