+++ test/sdk/test_import.py
@@ -1,17 +1,10 @@
-# pylint: disable=protected-access,missing-function-docstring, missing-class-docstring
-# pylint: disable=missing-module-docstring, missing-class-docstring
-# -*- coding: utf-8 -*-
-import sys
 from pathlib import Path
-sys.path.append(Path(__file__).parent.parent.absolute().name)
+import pytest_virtualenv as venv
 import sameproject.sdk as same
 import subprocess
-import sys
-
-import pytest
-import pytest_virtualenv as venv
-
 import importlib
+import pytest
+import sys
 
 
 @pytest.fixture
